refactor: CoinAmount 클래스 경로 변경
CoinAmount 클래스 경로를 Coin 과 같은 경로로 변경한다.
